requirements.txt
aiohttp>=3.8.0
asyncpg>=0.27
pandas>=2.0
pyarrow>=12.0
python-dotenv>=1.0
pydantic>=2.0
pydantic-settings>=2.0
alembic>=1.11
google-auth>=2.20
google-api-python-client>=2.95
gspread>=5.10
requests>=2.31
tenacity>=8.2
pytest>=7.0
pytest-asyncio>=0.21
aiofiles>=23.1
